requirements.txt
pydbus
rpi-gpio